Fixed dirty state so that saving does not prompt user
Before this fix the user would be prompted to save their work even if
they clicked the save button. This appears to be to do with the
asynchronous nature of the pushstate and set state calls. Now the
redirect is done as part of the callback after setting the dirty flag
as false.